requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
cryptography==41.0.5
dataclasses-json==0.6.1
decorator==5.1.1
Django==4.2.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
frozenlist==1.4.0
greenlet==3.0.1
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
idna==3.4
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.326
langsmith==0.0.54
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
Pillow==10.1.0
playsound==1.3.0
pycairo==1.25.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
PyJWT==2.8.0
pysqlite3==0.5.2
python-dotenv==1.0.0
python-telegram-bot==12.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
sniffio==1.3.0
SpeechRecognition==3.10.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tenacity==8.2.3
tornado==6.3.3
tqdm==4.66.1
twilio==8.10.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2